+++ src/model-function/generate-transcription/generateTranscription.ts
@@ -12,7 +12,7 @@ import {
  * @example
  * const data = await fs.promises.readFile("data/test.mp3");
  *
- * const transcription = await transcribe(
+ * const transcription = await generateTranscription(
  *   new OpenAITranscriptionModel({ model: "whisper-1" }),
  *   {
  *     type: "mp3",
@@ -20,14 +20,14 @@ import {
  *   }
  * );
  */
-export function transcribe<DATA>(
+export function generateTranscription<DATA>(
   model: TranscriptionModel<DATA, TranscriptionModelSettings>,
   data: DATA,
   options?: FunctionOptions
 ): ModelFunctionPromise<string> {
   return new ModelFunctionPromise(
     executeStandardCall({
-      functionType: "transcription",
+      functionType: "generate-transcription",
       input: data,
       model,
       options,